+++ core/src/main/java/com/graphhopper/storage/DataAccess.java
@@ -132,6 +132,6 @@ public interface DataAccess extends Storable<DataAccess>
      */
     DAType getType();
 
-    public void setTrfCnt( int edgeOrShortcutId, int time, int trafficCount );
-    public int getTrfCnt (int edgeId, int time);
+    public void setTrfCnt( int edgeOrShortcutId, int time, double trafficCount );
+    public double getTrfCnt (int edgeId, int time);
 }
